apps/photos/requirements.txt
Flask
python-jose
